@@ -3,7 +3,7 @@ description: Demonstrates how to use the scrollable_bottom_sheet plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
