@@ -1,11 +1,11 @@
 name: scrollable_bottom_sheet
 description: Simple Scrollable Persistent Bottom Sheet using GestureDetector
-version: 0.2.1
+version: 1.0.0
 author: Richardo <richardothayeb@gmail.com>
 homepage: https://github.com/ricnaaru/scrollable_bottom_sheet/tree/master
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
