@@ -33,7 +33,7 @@ abstract class ScrollableInterface {
 }
 
 class ScrollableController {
-  ScrollableInterface _interface;
+  ScrollableInterface? _interface;
 
   set interface(ScrollableInterface value) {
     _interface = value;
@@ -41,12 +41,12 @@ class ScrollableController {
 
   //re-set minimum height to make [ScrollableBottomSheetByContent] possible
   void setMinimumHeight(double newMinimumHeight) {
-    if (_interface != null) _interface.setMinimumHeight(newMinimumHeight);
+    if (_interface != null) _interface!.setMinimumHeight(newMinimumHeight);
   }
 
   //re-set half height to make [ScrollableBottomSheetByContent] possible
   void setHalfHeight(double newHalfHeight) {
-    if (_interface != null) _interface.setHalfHeight(newHalfHeight);
+    if (_interface != null) _interface!.setHalfHeight(newHalfHeight);
   }
 
   /// animate current Bottom Sheet to [minimumHeight] (if specified)
@@ -54,7 +54,7 @@ class ScrollableController {
   ///
   /// See [example/main.dart] for example
   void animateToMinimum(BuildContext context, {bool willPop = false}) {
-    if (_interface != null) _interface.animateToMinimum(context, willPop: willPop);
+    if (_interface != null) _interface!.animateToMinimum(context, willPop: willPop);
   }
 
   /// animate current Bottom Sheet to 0.0
@@ -62,7 +62,7 @@ class ScrollableController {
   ///
   /// See [example/main.dart] for example
   void animateToZero(BuildContext context, {bool willPop = false}) {
-    if (_interface != null) _interface.animateToZero(context, willPop: willPop);
+    if (_interface != null) _interface!.animateToZero(context, willPop: willPop);
   }
 
   /// animate current Bottom Sheet to initialHeight
@@ -70,7 +70,7 @@ class ScrollableController {
   ///
   /// See [example/main.dart] for example
   void animateToHalf(BuildContext context) {
-    if (_interface != null) _interface.animateToHalf(context);
+    if (_interface != null) _interface!.animateToHalf(context);
   }
 
   /// animate current Bottom Sheet to maximum height available
@@ -78,6 +78,6 @@ class ScrollableController {
   ///
   /// See [example/main.dart] for example
   void animateToFull(BuildContext context) {
-    if (_interface != null) _interface.animateToFull(context);
+    if (_interface != null) _interface!.animateToFull(context);
   }
 }
